receita_app/requirements.txt
beautifulsoup4>=4.9.3
bs4>=0.0.1
Flask>=2.3.0
greenlet>=1.1.0
importlib-metadata>=4.5.0
lxml>=4.6.3
numpy>=1.20.3
pandas>=1.2.4
psycopg2-binary>=2.9.1
python-dateutil>=2.8.1
python-dotenv==1.0.0
pytz>=2021.1
requests==2.30.0
six>=1.16.0
soupsieve>=2.2.1
SQLAlchemy>=1.4.18
typing-extensions>=3.10.0.0
tzdata==2023.3
urllib3==2.0.2
wget>=3.2
zipp>=3.4.1
